orders/src/events/listeners/ticket-created-listener.ts
import { Subjects, TicketCreatedEvent, Listener } from "@msetick/common";
import { Message } from "node-nats-streaming";
import { Ticket } from "../../models/ticket";
import { queueGroupName } from "./queue-group-name";

export class TicketCreatedListener extends Listener<TicketCreatedEvent> {
    subject: Subjects.TicketCreated = Subjects.TicketCreated;
    queueGroupName = queueGroupName;
    async onMessage(data: TicketCreatedEvent["data"], msg: Message) {
        console.log("Event data!", data);
        const { id, title, price } = data;
        const ticket = Ticket.build({
            id,
            title,
            price,
        });
        await ticket.save();
        console.log("ticket saved");
        msg.ack();
    }
}
